Moved favorites to UserProfile component
